Update credentials.py
I assume we would like to check that a token has expired and then refresh it if it has, however when checking expiry the `self.oauth_expires_at` is a string so gives a `TypeError`.  This PR converts the `self.oauth_expires_at` back to a date so `return self.oauth_expires_at <= (now + datetime.timedelta(seconds=CONSERVATIVE_SECONDS))` doesn't fail.